requirements-minimal.txt
# Python 3.10+ only. Smaller set to run start.py, reject/intent infer, and chatnlu_infer.
# Install with: .\.venv\Scripts\python.exe -m pip install -r requirements-minimal.txt

flask>=3.0,<4
flask-socketio>=5.3,<6
python-socketio>=5.11.0
python-engineio>=4.9.0
werkzeug>=3.0,<4
redis>=4.5
requests>=2.28
python-dotenv>=1.0
fastapi>=0.109
uvicorn[standard]>=0.27
numpy>=1.24
scikit-learn>=1.3
torch>=2.0
